requirements.txt
# Web Framework
fastapi==0.115.10
uvicorn==0.34.0
python-multipart==0.0.20

# Database
motor==3.7.0
beanie==1.29.0
pymongo==4.11.1
qdrant-client==1.13.2

# AI/ML
openai==1.65.1

# Web Scraping
beautifulsoup4==4.13.3
requests==2.32.3

# Utils
python-dotenv==1.0.1
loguru==0.7.3
pydantic-core==2.31.1
typing-extensions==4.12.2

# Testing and Development
pytest==8.3.4
pytest-asyncio==0.25.3

# Linters
black==25.1.0
isort==6.0.1